📚 Update README.md
Made changes to the README.md file to provide a more comprehensive and up-to-date documentation. Added instructions on how to install and run the project, along with details about its features and dependencies. Also included information about contributing to the project and guidelines for submitting bug reports and feature requests.